(#2) Establish HTTPS for the server.
Add https and koa-sslify to the dependencies.
Add a new configuration file for the SSL part.
Start a HTTPS server next to the existing HTTP one.
Force HTTTPS connections with koa-sslify, which redirects clients
automatically.
Establish a bootstrap script.
This provide so far a function to check the SSL setup.
For the development environment it does generate the SSL key and
certificate automatically.