(#2) Establish logging for the server.
Add bunyan and koa-bunyan-logger to the dependencies.
Create a new configuration file for the logging part.
Define the logger script which initiate the logger.
Add the logger instance to koa application instance.
Log the connections of the server.
Extend the bootstrap script to initialize the logging.
Add some logging commands on starting the server.
Add the folder for the logs to the ignore list.