(#2) Implement the first simple router for export.
Add koa-router to the dependencies.
Implement the export API in a router module.
Establish a dummy parameter check function for the permit ID.
Establish the GET function for the API permitXMLExport function.
Add freezed objects as enumerations for the route names and arguments.
The router works and simply return a plain text for each permit ID.