(#2) Implement the connection to the blockchain.
Add web3 to the dependencies to be able connecting to the ethereum blockchain.
Establish a property file for the contracts, which read configuration
data as the abi description and the contract address.
Both have to be provided by the truffle migration and are not part of
the sources, so the whole data folder has been added to be ignored.
A helper script for this process will follow.
Add new constants file for contract names.
Extend the general properties to define the web3 provider.
Implement the contract reader to read data from the blockchain.
Adjust the export router to try getting the permit from the chain.
So far only a negative test was possible, cause the DApp UI does not
provide the whole permit ID.