(#2) Introduce the XML conversion of permits.
Switch back to a stable Web3 package version and adjust the usage for that.
Move the contract reader to the utility folder so its not a single
folder situation with the 'api' directory.
Add xml-js to the dependencies to generate XMLs from JSONs.
Introduce the xml_converter which offers the convertPermitToXml
function.
Add new constants for the XML tags.
Adjust the export router to use the converter.