(#2) Introduce the conversion cache.
Add the cache handler as new class.
Define new constants file for all provided conversion types.
Implement the functionality to store a permit for a specific conversion
type to the cache and to retrieve a permits all cached files.
Extend the export router to use the cache.
Extend the bootstrap script for the caching.
Unify the argument names and extend documentation in the contract
reader.
Define the folder where to cache in the general properties.