Extend XML export functionality.
The contract reader has been extend to read specimens as well.
The converter can handle permits and specimens now.
The conversion of specimens is reused for the permits.
Establish work around to handle embedded specimens in permits or single
exported.
Extend the XML property keys for that purpose.
The specimen export is not used by the API so far.
This implementation base on PR #77 in the DApp repository.