Extend export router interface for specimens.
Add argument check function for the specimen identifier.
Add export function for specimens.
Caching for specimens is not supported so far.
Extend the constants route names and arguments.
Add new constants for the route paths.
Add (more) documentation to the converter class.
New naming convention in the converter, to differ between converting
into file types and restructure a blockchain permit.