Establish caching for all objects.
Refactor the cache handler to be able to use it for all defined
blockchain objects.
Introduce the blockchain object constants for that purpose.
Adjust the export router to use the caching for speciemens and fix the
existing calls.
Remove the cache bootstrap function, cause the fancy function to create
subfolders for all objects is able to handle paths with endless folder
hierarchy.