Make the whole thing much more generic.
Use only one route only.
This route is used to export and convert any blockchain object that is
provided as first path parameter and in the constants.
Extend the rest of the code which wasn't generic before to work on this
blockchain objects constants as well, so everything works fine with
this.
By this approach the application is high flexible and extendable.