(#2) Add optional query parameter for conversion type.
The route to export blockchain objects, can handle a query parameter
called 'conversion', to specify the data type the object should be
converted after the export.
The parameter value will be validated if it is a supported type.
If no query is given, a default type (XML) is used.
By this the interface stays stable.