(#8) Establish auto-formatting on git commits.
Add ESLinter to the dependencies as well as some plugins for that.
Add Prettier to the dependencies and use it as fix tool for the ESLinter.
Establish git-hooks by adding Husky to the dependencies.
Define `precommit` hook which calls lint-staged (also added to the
dependencies).
Add `postcommit` hook to restage adjusted files by autofix format.